refactor: auth와 user 관련 파일 분리
- auth 라우터와 관련된 파일을 auth 디렉토리로 이동
- user 라우터와 관련된 파일을 user 디렉토리로 이동
